Feat: 게시글에 댓글 목록과 좋아요 숫자를 보이도록 하는 기능 추가 (#43)
* Fix: GolbalException이 에러 메세지가 아닌 에러를 출력하던 오류 수정

* Chore: RequestMapping의 경로 수정

매핑이 보다 직관적으로 보이도록 메서드 매핑에 해당 엔티티 이름이 포함되도록 변경

* Fix: 엔티티 이름이 지정되지 않아 테이블이 생성되지 않던 오류 수정

엔티티 네이밍을 통해 해결

* Fix: 댓글에 빈 칸을 적어도 댓글이 작성되던 오류 수정

@NotNull 어노테이션을 content에 적용해 해결

* Feat: 게시글에 댓글 목록과 좋아요 숫자를 보이도록 하는 기능 추가

추가적으로 댓글의 좋아요 숫자도 보이도록 하는 기능 추가

+++ src/main/java/com/sparta/instagramclonebe/domain/like/repository/PostLikeRepository.java
@@ -7,9 +7,7 @@ import java.util.Optional;
 
 public interface PostLikeRepository extends JpaRepository<PostLike,Long> {
     Optional<PostLike> findByPostIdAndUserId(Long id, Long id1);
-
     void deleteByPostIdAndUserId(Long id, Long id1);
-
-    int countPostLikeByPostId(Long id);
+    Long countPostLikeByPostId(Long id);
 
 }
